requirements.txt
wheel
alembic==1.0.11
asn1crypto==1.3.0
attrs==19.3.0
blinker==1.4
certifi==2020.6.20
cffi==1.12.3
chardet==3.0.4
clg==2.0.0
Click==7.0
colorama==0.4.3
configparser==5.0.0
config2
crayons==0.3.1
cryptography==2.7
cx_Oracle==7.3.0
dnspython==1.16.0
docopt==0.6.2
dominate==2.3.5
email-validator==1.0.4
enum34==1.1.6
Flask==1.0.3
Flask-Admin==1.5.3
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.5.2
Flask-MySQL==1.4.0
flask-nav==0.6
Flask-SQLAlchemy==2.4.3
fpdf==1.7.2
idna==2.8
importlib-metadata==1.7.0
is-safe-url==1.0
itsdangerous==1.1.0
Jinja2==2.10.1
Mako==1.0.13
MarkupSafe==1.1.1
migrant==1.3.1
more-itertools==8.4.0
packaging==20.4
pathtools==0.1.2
peewee==3.9.6
peewee-migrations==0.3.18
pluggy==0.13.1
py==1.9.0
pycparser==2.19
PyMySQL==0.9.3
pyodbc==4.0.30
pyparsing==2.4.7
pytest==5.4.3
pytest-base-url==1.4.2
pytest-html==2.1.1
pytest-metadata==1.10.0
pytest-selenium==1.17.0
pytest-variables==1.9.0
pytest-watch==4.2.0
python-dateutil==2.8.0
python-editor==1.0.4
PyYAML==5.3
requests==2.24.0
selenium==3.141.0
six==1.12.0
SQLAlchemy==1.3.18
termcolor==1.1.0
urllib3==1.25.9
visitor==0.1.3
watchdog==2.1.2
wcwidth==0.2.5
webdriver-manager==3.2.1
Werkzeug==0.15.5
wtf-peewee==3.0.0
WTForms==2.2.1
XlsxWriter==1.1.8
zipp==3.1.0
